+++ package.json
@@ -20,8 +20,9 @@
     "next-svgr": "^0.0.2",
     "qs": "^6.13.1",
     "react": "^18.3.1",
-    "react-beautiful-dnd": "^13.1.1",
     "react-calendar": "^5.1.0",
+    "react-dnd": "^16.0.1",
+    "react-dnd-html5-backend": "^16.0.1",
     "react-dom": "^18.0.0",
     "react-dropzone": "^14.3.5",
     "react-slick": "^0.30.3",
@@ -34,7 +35,6 @@
     "@types/node": "^20",
     "@types/qs": "^6.9.17",
     "@types/react": "^18",
-    "@types/react-beautiful-dnd": "^13.1.8",
     "@types/react-dom": "^18",
     "@types/react-slick": "^0.23.13",
     "eslint": "^8",
