+++ src/app/(creator)/creator/_component/OptionButton.tsx
@@ -20,9 +20,9 @@ const OptionButton: React.FC<OptionButtonInfo> = ({ items }) => {
           className="absolute right-0 top-9 z-30 w-[110px] h-fit flex flex-col rounded-[10px] border-gray-400/20 border p-[4px] bg-white"
           style={{ boxShadow: '0px 0px 4px 0px rgba(164, 164, 164, 0.12)' }}
         >
-          {items.map((item, i) => (
+          {items.map((item) => (
             <button
-              key={i}
+              key={item.text}
               className="hover:bg-brand-gray rounded-[6px] p-[5px] text-[13px] flex w-full"
               onClick={() => {
                 item.onClick(), setOption(false);
